### 2017-11-29 #474 @whedon
[PRE REVIEW]: The Experiment Factory: Reproducible Experiment Containers
**Submitting author:** @vsoch (<a href="http://orcid.org/0000-0002-4387-3819">Vanessa Sochat</a>)
**Repository:** <a href="https://www.github.com/expfactory/expfactory" target ="_blank">https://www.github.com/expfactory/expfactory</a>
**Version:** v3.0.0
