### 2017-11-30 7689fc5
fix typo in paper.md
