### 2017-12-03 b65e5c4
testing adding randomize option
