### 2017-12-04 f99d0ba
modified:   docs/pages/1-generate.md 	modified:   docs/pages/2-usage.md 	modified:   expfactory/server.py 	modified:   expfactory/templates/build/docker/startscript.sh
