### 2017-12-04 d7900d9
adding ability to select specific experiments in advance
