+++ expfactory/templates/build/docker/startscript.sh
@@ -27,7 +27,8 @@ usage () {
 
                 --studyid:  specify a studyid to override the default
                 --randomize: select experiment order at random
-                --no-randomize: select experiment order manually
+                --no-randomize: select experiment order manually in the GUI
+                --experiments: a comma separated list of experiments (manual ordering) 
 
          Examples:
 
@@ -65,6 +66,13 @@ while true; do
             env | grep ${1:-}
             exit
         ;;
+        -e|experiments|--experiments)
+            shift
+            EXPFACTORY_EXPERIMENTS="${1:-}"
+            shift
+            export EXPFACTORY_EXPERIMENTS
+            exit
+        ;;
         --database|--db)
             shift
             EXPFACTORY_DATABASE=${1:-}
@@ -87,13 +95,11 @@ while true; do
             shift
             EXPFACTORY_RANDOM="true"
             export EXPFACTORY_RANDOM
-            shift
         ;;
         --no-randomize)
             shift
             EXPFACTORY_RANDOM="false"
             export EXPFACTORY_RANDOM
-            shift
         ;;
         --lib)
             echo "Experiments in the library:"
