+++ expfactory/experiment.py
@@ -101,8 +101,8 @@ def get_selection(available, selection, base='/scif/apps'):
     '''we compare the basename (the exp_id) of the selection and available, 
        regardless of parent directories'''
 
-    if isinstance(selection,str):
-        selection = selection.split(' ')
+    if isinstance(selection, str):
+        selection = selection.split(',')
 
     available = [os.path.basename(x) for x in available]
     selection = [os.path.basename(x) for x in selection]
