+++ expfactory/database/relational.py
@@ -90,7 +90,7 @@ def save_data(self,session, exp_id, content):
         if "data" in content:
             content = content['data']
 
-        result = Result(data=data,
+        result = Result(data=content,
                         exp_id=exp_id,
                         participant_id=p.id) # check if changes from str/int
         self.session.add(result)
