+++ expfactory/database/relational.py
@@ -46,6 +46,7 @@ from expfactory.defaults import (
 )
 from glob import glob
 import os
+import pickle
 import json
 import sys
 
@@ -90,6 +91,8 @@ def save_data(self,session, exp_id, content):
         if "data" in content:
             content = content['data']
 
+        print(content)
+        pickle.dump(content. open('/data/content.pkl','wb'))
         result = Result(data=content,
                         exp_id=exp_id,
                         participant_id=p.id) # check if changes from str/int
