### 2017-12-05 dde1d8f
updating docs about sqlite
