### 2017-12-05 4336327
removing exit where it shouldn't be!
