### 2017-12-05 be5726c
updating docs on ordering
