### 2017-12-05 #474 @vsoch
oh, of course! I think a mix of linux, python, and general experimental knowledge (e.g., psychology, sociology, or economics) would be a logical match, so I'll recommend any of @cMadan @nirum @oesteban  @jsta and @ctb. If we need additional recommendations I could look into others outside of the provided list, please let me know.
