### 2017-12-06 19311d5
adding start to headless mode, first will test generation of ids.
