+++ expfactory/server.py
@@ -30,7 +30,8 @@ from expfactory.experiment import (
 from expfactory.database import (
     init_db, 
     generate_subid,
-    save_data
+    save_data,
+    generate_user
 )
 from flask import Flask
 from flask_wtf.csrf import (
@@ -131,6 +132,7 @@ class EFServer(Flask):
 EFServer.init_db = init_db
 EFServer.save_data = save_data
 EFServer.generate_subid = generate_subid
+EFserver.generate_user = generate_user
 app = EFServer(__name__)
 app.config.from_object('expfactory.config')
 
