+++ expfactory/database/models.py
@@ -42,15 +42,17 @@ from sqlalchemy import (
     func
 )
 from sqlalchemy.orm import relationship, backref
+from uuid import uuid4
 from expfactory.database import Base
 
 
 class Participant(Base):
-    '''A participant in a local assessment. Names must be unique
+    '''A participant in a local assessment. id must be unique.
     '''
     __tablename__ = 'participant'
     id = Column(Integer, primary_key=True)
     name = Column(String(150))
+    token = Column(String(36))
     results = relationship('Result', lazy='select',
                            backref=backref('participant', lazy='joined'))
     def __init__(self, name=None):
@@ -59,6 +61,9 @@ class Participant(Base):
     def __repr__(self):
         return '<Participant %r>' % (self.name)
 
+    def url(self):
+        '''return the participant url'''
+        
 
 class Result(Base):
     '''a result is an experiment name, json dump, and datetime'''
