+++ expfactory/database/__init__.py
@@ -7,3 +7,6 @@ else:
     from .relational import *
     if EXPFACTORY_DATABASE.startswith('sqlite'):
         from .sqlite import *
+
+# Shared functions across database types
+from shared import *
