### 2017-12-06 015427d
removing shared function
