### 2017-12-06 #474 @cMadan
This looks like a great project, but unfortunately I don't have time to review it currently.
