+++ expfactory/views/__init__.py
@@ -1,2 +1,4 @@
+from .general import *
+from .headless import *
 from .main import *
 from .database import *
