### 2017-12-06 8fc89ce
adding headless to server settings
