+++ expfactory/views/__init__.py
@@ -1,4 +1,3 @@
-from .general import *
-from .headless import *
 from .main import *
 from .database import *
+from .headless import *
