+++ expfactory/views/main.py
@@ -52,6 +52,7 @@ from expfactory.views.utils import (
 )
 from expfactory.server import app
 from .general import *
+from .headless import *
 
 from random import choice
 import logging
