### 2017-12-06 6e99093
fixing portal headless page
