### 2017-12-06 9ea6c86
need to check for fullpath
