### 2017-12-06 d56f2d3
headless shouldn't need to make numbered folders
