### 2017-12-06 af1c2c4
bug with function to create user
