### 2017-12-06 7d07e73
modified:   MANIFEST.in
