+++ MANIFEST.in
@@ -1,4 +1,4 @@
 include README.md LICENSE
 recursive-include expfactory *
 recursive-include expfactory * *
-global-exclude docs * *
+global-exclude docs
