### 2017-12-06 dd8d0eb
use token to determine if create uuid folder
