### 2017-12-06 1b6b1a0
need to redirect to next if have token
