### 2017-12-06 a12cbc1
adding print lines to debug
