### 2017-12-07 c70c76e
adding session to server
