+++ expfactory/views/main.py
@@ -95,6 +95,7 @@ def home():
     if app.headless:
         if "token" not in session:
             form = EntryForm()
+            session['experiments'] = app.experiments # list
             return render_template('routes/entry.html', form=form)
         return redirect('/next')
 
