### 2017-12-07 22f31aa
username cannot be None
