### 2017-12-07 488c1cd
fixing up return_variables
