### 2017-12-07 b56d7c1
modified:   docs/pages/2-usage.md 	modified:   expfactory/cli/users.py
