### 2017-12-08 a6ea812
modified:   expfactory/cli/users.py 	modified:   expfactory/database/relational.py
