+++ expfactory/database/relational.py
@@ -62,7 +62,7 @@ import sys
 #
 ################################################################################
 
-def generate_subid(self, token=None, digits=5):
+def generate_subid(self, token=None, digits=5, return_user=False):
     '''generate a new user in the database, still session based so we
        create a new identifier.
     '''    
@@ -74,6 +74,8 @@ def generate_subid(self, token=None, digits=5):
     self.session.add(p)
     self.session.commit()
     print('Session Participant id: %s' % p.id)
+    if return_user is True:
+        return p
     return p.id
 
 
@@ -82,11 +84,9 @@ def generate_user(self, digits=5):
        create a new identifier. This function is called from the users new 
        entrypoint, and it assumes we want a user generated with a token.
     '''
-    from expfactory.database.models import Participant
     token = str(uuid.uuid4())
-    subid = self.generate_subid(digits=digits, token=token)
-    print('RELATIONAL: generating user %s' %subid)
-    return subid
+    user = self.generate_subid(digits=digits, token=token, return_user=True)
+    return "%s\t%s" %(user.id, user.token)
 
 
 def validate_token(self, token):
@@ -111,9 +111,7 @@ def list_users(self):
     participants = Participant.query.all()
     users = []
     for participant in participants:
-        subid = participant.id
-        if participant.token is not None:
-            subid = participant.token
+        subid = "%s\t%s" %(participant.id, participant.token)
         users.append(subid)
     return users
 
