### 2017-12-08 e2c93bb
updating print function for single user
