+++ expfactory/server.py
@@ -31,6 +31,7 @@ from expfactory.database import (
     init_db, 
     list_users,
     generate_subid,
+    print_user,
     save_data,
     generate_user,
     validate_token
@@ -139,6 +140,7 @@ EFServer.generate_subid = generate_subid
 EFServer.generate_user = generate_user
 EFServer.validate_token = validate_token
 EFServer.list_users = list_users
+EFServer.print_user = print_users
 app = EFServer(__name__)
 app.config.from_object('expfactory.config')
 
