### 2017-12-08 9b30563
updating print function for single user
