### 2017-12-08 0ae0a23
adding functions to finish, restart, etc.
