### 2017-12-08 11dfbbf
updating print function for single user
