+++ expfactory/cli/__init__.py
@@ -52,7 +52,7 @@ def get_parser():
 
     # Users manager
     users = subparsers.add_parser("users",
-                                     help="Manager for interacting with users")
+                                   help="Manager for interacting with users")
     users.add_argument('--new', dest="new",
                        help="generate new user tokens, recommended for headless runtime.",
                        default=None, type=int)
