### 2017-12-08 134cc77
subparser should print help
