+++ expfactory/views/general.py
@@ -48,18 +48,6 @@ import os
 import json
 
 
-# EXPERIMENT ROUTER ############################################################
-
-def router():
-
-    # Redirects to another template view
-    experiment = app.get_next(session)
- 
-    if experiment is not None:
-        app.logger.info('Next experiment is %s' % experiment)
-    return perform_checks('/experiments/%s' % experiment, do_redirect=True,
-                          next=experiment)
-
 
 # Home portal to start experiments
 def portal():
