### 2017-12-08 ff992a9
subparser should print help
