### 2017-12-08 4ea3e21
adding debugging info
