+++ expfactory/views/main.py
@@ -114,6 +114,7 @@ def save():
     '''
     if request.method == 'POST':
         exp_id = session.get('exp_id')
+        app.logger.debug('Saving data for %s' %exp_id)
 
         fields = get_post_fields(request)
         result_file = app.save_data(session=session, content=fields, exp_id=exp_id)
