+++ expfactory/views/main.py
@@ -137,10 +137,10 @@ def next():
 
     # To generate redirect to experiment
     experiment = app.get_next(session)
-    app.logger("Selected experiment is %s" %experiment)
+    app.logger.debug("Selected experiment is %s" %experiment)
  
     if experiment is not None:
-        app.logger.info('Next experiment is %s' % experiment)
+        app.logger.debug('Next experiment is %s' % experiment)
         return perform_checks('/experiments/%s' % experiment, do_redirect=True,
                               next=experiment)
 
