### 2017-12-08 dcad601
forgot to add logs file
