### 2017-12-08 44222de
need to print user on generation
