+++ docs/pages/2-usage.md
@@ -81,95 +81,19 @@ If you ask for non random order without giving a list, you will present the expe
 
 
 ### Generate tokens
-A "token" is basically a subject id that is intended to be used once, and can be sent securely to your participants to access the experiments. You will need to generate them, and we can use `exec` to execute a command to the container to do this. If we just run the general users command, we learn better how to use it:
+A "token" is basically a subject id that is intended to be used once, and can be sent securely to your participants to access the experiments. The token can be refreshed, revoked, or active. You will need to generate them, and briefly it looks like this:
+
 
 ```
-docker exec experiments expfactory users
-See expfactory users --help for usage
-```
-
-The full usage:
-
-```
-usage: expfactory users [-h] [--new NEW] [--list] [--revoke REVOKE]
-                        [--refresh REFRESH] [--restart RESTART]
-                        [--finish FINISH]
-
-optional arguments:
-  -h, --help         show this help message and exit
-  --new NEW          generate new user tokens, recommended for headless
-                     runtime.
-  --list             list current tokens, for a headless install
-  --revoke REVOKE    revoke token for a user id, ending the experiments
-  --refresh REFRESH  refresh a token for a user
-  --restart RESTART  restart a user, revoking and then refresing the token
-  --finish FINISH    finish a user session by removing the token
-```
-
-Let's try creating three new users with the `--new` flag:
-
-```
+docker exec experiments expfactory users --help
 docker exec experiments expfactory users --new 3
-DATABASE	TOKEN
-/scif/data/expfactory/41a451cc-7416-4fab-9247-59b1d65e33a2	41a451cc-7416-4fab-9247-59b1d65e33a2[active]
-/scif/data/expfactory/6afabdd5-7d5e-48dc-a3b2-ade235d2e0a6	6afabdd5-7d5e-48dc-a3b2-ade235d2e0a6[active]
-/scif/data/expfactory/3251fd0e-ba3e-4089-b01a-28dfa03f1fbd	3251fd0e-ba3e-4089-b01a-28dfa03f1fbd[active]
 ```
 
-The result here will depend on the database type. 
-
- - `DATABASE`: The above shows a filesystem save, so a `DATABASE` refers to the folder, and remember this is internal to the container, so you might have `/scif/data` mapped to a different folder on your host. A relational database would have the `DATABASE` column correspond with the index. 
- - `TOKEN`: The token corresponds with the folder on the filesystem, and shown also is the participant status (e.g., `active`).
-
-You can copy paste this output from the terminal, or pipe into a file instead:
-
-```
-docker exec experiments expfactory users --new 3 >> participants.tsv
-```
-
-You can also issue these commands by shelling inside the container, which we will do for the remainder of the examples:
-
-```
-docker exec -it experiments bash
-```
-
-If you ever need to list the tokens you've generated, you can use the `users --list` command. Be careful that the environment variable `EXPFACTORY_DATABASE` is set to be the one that you intend. For example, a filesystem database setting will print all folders found in the mapped folder given this variable is set to `filesystem`. In the example below, we list users saved as folders on the filesystem, and note that the command can be used for headless (token-named) users as well as traditional.
-
-Headless experiment runs have tokens for identifiers
-
-```
- expfactory users --list
-DATABASE	TOKEN
-/scif/data/expfactory/41a451cc-7416-4fab-9247-59b1d65e33a2	41a451cc-7416-4fab-9247-59b1d65e33a2[active]
-/scif/data/expfactory/6afabdd5-7d5e-48dc-a3b2-ade235d2e0a6	6afabdd5-7d5e-48dc-a3b2-ade235d2e0a6[active]
-/scif/data/expfactory/3251fd0e-ba3e-4089-b01a-28dfa03f1fbd	3251fd0e-ba3e-4089-b01a-28dfa03f1fbd[active]
[…]
