+++ expfactory/database/relational.py
@@ -62,7 +62,7 @@ import sys
 #
 ################################################################################
 
-def generate_subid(self, token=None, digits=5, return_user=False):
+def generate_subid(self, token=None, return_user=False):
     '''generate a new user in the database, still session based so we
        create a new identifier.
     '''    
@@ -114,13 +114,13 @@ def list_users(self, user=None):
 
 # Actions ######################################################################
 
-def generate_user(self, digits=5):
+def generate_user(self):
     '''generate a new user in the database, still session based so we
        create a new identifier. This function is called from the users new 
        entrypoint, and it assumes we want a user generated with a token.
     '''
     token = str(uuid.uuid4())
-    return self.generate_subid(digits=digits, token=token, return_user=True)
+    return self.generate_subid(token=token, return_user=True)
 
 
 def finish_user(self, subid):
