+++ expfactory/database/filesystem.py
@@ -150,7 +150,6 @@ def restart_user(self, subid):
             if os.path.exists(folder):
                 os.rename(folder, data_base)
                 self.logger.info('Restarting %s, folder is %s.' % (subid, data_base))
-                return folder
 
         self.logger.warning('%s does not have revoked or finished folder, no changes necessary.' % (subid))
         return data_base    
@@ -178,9 +177,9 @@ def refresh_token(self, subid):
     '''refresh or generate a new token for a user. If the user is finished,
        this will also make the folder available again for using.'''
     if os.path.exists(self.database):    # /scif/data/<study_id>
-        data_base = "%s/%s" %(self.data_base, subid)
+        data_base = "%s/%s" %(self.database, subid)
         if os.path.exists(data_base):
-            refreshed = "%s/%s" %(self.data_base, str(uuid.uuid4()))
+            refreshed = "%s/%s" %(data_base, str(uuid.uuid4()))
             os.rename(data_base, refreshed)
             return renamed
         self.logger.warning('%s does not exist, cannot rename %s' % (data_base, subid))
