### 2017-12-08 e8746ea
modified:   docs/pages/2-usage.md 	modified:   expfactory/cli/users.py
