### 2017-12-08 7ee3ad8
new file:   templates/build/docker/builder-dev/Dockerfile 	new file:   templates/build/docker/builder-dev/README.md 	new file:   templates/build/docker/builder-dev/entrypoint.sh
