+++ expfactory/templates/build/docker/builder-dev/README.md
@@ -14,6 +14,7 @@ build the image
 
 ```
 docker build -t vanessa/expfactory-builder-dev .
+docker push vanessa/expfactory-builder-dev
 ```
 
 Follow instructions in the [main builder](../builder) README for using the builder. If you have questions please [get in touch](https://www.github.com/expfactory/issues).
