+++ expfactory/templates/build/docker/builder-dev/Dockerfile
@@ -1,4 +1,4 @@
-FROM ubuntu:14.04
+FROM ubuntu:16.04
 
 # docker build -t vanessa/expfactory-builder-dev .
 
