### 2017-12-11 90243aa
fixing bug with experiment finish
