+++ docs/pages/3-contribute.md
@@ -156,6 +156,14 @@ For the cases above, you can use the `vanessa/expfactory-builder` image to run t
 
 
 ### Test an Experiment
+Testing an experiment primarily means two things: some kind of static testing for content, and a runtime test that the experiment functions as we would expect.
+
+### Runtime Test
+When you submit an experiment for review, given that the repository for the experiment is also hosting it on the Github pages associated with the repository, it's likely easy enough for you and your reviewers to test the experiment using Github pages. However, for predictable experiment layouts (e.g., jspsych) we have developed an experiment robot that can run through the experiment for you. We provide the robot as a container, with source code in the [examples](https://github.com/expfactory/expfactory/tree/master/examples) folder.
+
+TODO write me
+
+### Static Testing
 You have two options to test experiments on your host using `vanessa/expfactory-builder`. If you want to test a single experiment, meaning a folder with a `config.json` file:
 
 ```
@@ -212,6 +220,7 @@ Found experiment emotion-regulation
 ```
 
 Remember that these tests are primarily looking at metadata, and runtime of your experiment still will need to be tested by a human, primarily when installed in the container.
+
 
 ### Test a Contribution
 This set of tests is more stringent in that the test starts with one of more submissions (markdown files that you will ask to be added to the `_library` folder via a pull request) and goes through Github cloning to testing of your preview. Specifically it includes:
