### 2017-12-11 1f76010
updating entrypoint message
