### 2017-12-11 2ca2e73
finishing up fixes for filesystem
