### 2017-12-11 d4724f1
checking args.database should honor database selectionx
