### 2017-12-14 420c146
adding first test for builder on circle (need to connect)
