### 2017-12-14 8df8955
more tweaks to testing
