+++ circle.yml
@@ -14,8 +14,7 @@ dependencies:
     - echo "The Experiment Factory!"
     - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
     - /bin/bash Miniconda3-latest-Linux-x86_64.sh -b 
-    - cd /tmp && git clone https://github.com/expfactory/expfactory.git
-    - cd /tmp/expfactory && $HOME/miniconda3/bin/python setup.py install
+    - cd $HOME/expfactory && $HOME/miniconda3/bin/python setup.py install
 
 test:
   override:
