+++ circle.yml
@@ -9,7 +9,6 @@ machine:
 dependencies:
   override:
     - docker info
-    - cd expfactory/templates/build/docker/builder && docker build --rm=false -t expfactory/expfactory-builder .
   pre:
     - echo "The Experiment Factory!"
     - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
