### 2017-12-14 2a3fef5
adding continuous integration builder
