+++ expfactory/testing/test_builder.sh
@@ -2,7 +2,7 @@
 
 echo "Testing builder writing to /tmp"
 mkdir -p /tmp/data
-docker run -v /tmp/data:/data vanessa/expfactory-builder build test-task
+docker run -v /tmp/data:/data vanessa/expfactory-builder-ci build test-task
 
 echo "Building container"
 cd /tmp/data
