+++ circle.yml
@@ -9,6 +9,7 @@ machine:
 dependencies:
   override:
     - docker info
+    - cd expfactory/templates/build/docker/builder-ci && /bin/bash build.sh
   pre:
     - echo "The Experiment Factory!"
     - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
