+++ expfactory/templates/build/docker/builder-ci/build.sh
@@ -1,6 +1,7 @@
 #!/use/bin/bash
 
 echo "Building CI builder for testing current PR..."
-cp -R ../../../../../expfactory $PWD
+cp -r ../../../../../expfactory/. $PWD/
+cp ../builder/entrypoint.sh $PWD
 ls
 docker build --rm=false -t vanessa/expfactory-builder-ci .
