+++ expfactory/templates/build/docker/builder-ci/Dockerfile
@@ -2,7 +2,7 @@ FROM ubuntu:16.04
 
 # docker build -t vanessa/expfactory-builder-ci .
 
-RUN apt-get update && apt-get install -y nginx git python3-pip python3-dev python libyaml-dev
+RUN apt-get update && apt-get install -y nginx git python3-pip python3-dev python apt-utils
 ENV DEBIAN_FRONTEND noninteractive
 
 ###################################
@@ -11,9 +11,8 @@ ENV DEBIAN_FRONTEND noninteractive
 
 WORKDIR /opt
 ADD . /opt
-WORKDIR /opt/expfactory
+RUN ls
 RUN python3 setup.py install
-RUN python3 -m pip install pyaml
 RUN mv /opt/entrypoint.sh /entrypoint.sh
 RUN chmod u+x /entrypoint.sh
 RUN mkdir -p /scif/apps
