+++ expfactory/templates/build/docker/builder-ci/build.sh
@@ -1,7 +1,7 @@
 #!/use/bin/bash
 
 echo "Building CI builder for testing current PR..."
-cp -r ../../../../../expfactory/. $PWD/
+cp -r ../../../../../. $PWD/
 cp ../builder/entrypoint.sh $PWD
 ls
 docker build --rm=false -t vanessa/expfactory-builder-ci .
