+++ expfactory/templates/build/docker/builder-ci/build.sh
@@ -1,7 +1,8 @@
 #!/use/bin/bash
 
 echo "Building CI builder for testing current PR..."
-cp -r ../../../../../. $PWD/
+ls ../../../../../
+cp -R ../../../../../. $PWD
 cp ../builder/entrypoint.sh $PWD
 ls
 docker build --rm=false -t vanessa/expfactory-builder-ci .
