+++ expfactory/templates/build/docker/builder-ci/build.sh
@@ -1,8 +1,7 @@
 #!/use/bin/bash
 
 echo "Building CI builder for testing current PR..."
-ls ../../../../../
-cp -R ../../../../../. $PWD
-cp ../builder/entrypoint.sh $PWD
-ls
-docker build --rm=false -t vanessa/expfactory-builder-ci .
+mkdir -p /tmp/data
+cp ../builder/entrypoint.sh /tmp/data
+cp Dockerfile /tmp/data
+cd /tmp/data && docker build --rm=false -t vanessa/expfactory-builder-ci .
