### 2017-12-14 1d8e257
adding build setup step
