+++ expfactory/templates/build/docker/builder-ci/build.sh
@@ -1,7 +1,4 @@
 #!/use/bin/bash
 
 echo "Building CI builder for testing current PR..."
-mkdir -p /tmp/data
-cp ../builder/entrypoint.sh /tmp/data
-cp Dockerfile /tmp/data
-cd /tmp/data && docker build --rm=false -t vanessa/expfactory-builder-ci .
+docker build --rm=false -t vanessa/expfactory-builder-ci .
