+++ circle.yml
@@ -10,7 +10,7 @@ dependencies:
   override:
     - docker info
     - cd expfactory/templates/build/docker/builder-ci && /bin/bash setup_build.sh
-    - /bin/bash /tmp/builder/build.sh
+    - cd /tmp/builder && /bin/bash build.sh
   pre:
     - echo "The Experiment Factory!"
     - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
