+++ expfactory/templates/build/docker/builder-ci/Dockerfile
@@ -12,6 +12,8 @@ ENV DEBIAN_FRONTEND noninteractive
 WORKDIR /opt
 ADD . /opt
 RUN ls
+WORKDIR /opt/expfactory
+RUN ls
 RUN python3 setup.py install
 RUN mv /opt/entrypoint.sh /entrypoint.sh
 RUN chmod u+x /entrypoint.sh
