+++ expfactory/testing/test_builder.sh
@@ -1,9 +1,8 @@
 #!/bin/sh
 
 echo "Testing builder writing to /tmp"
-mkdir -p /tmp/data
-docker run -v /tmp/data:/data vanessa/expfactory-builder-ci build test-task
+mkdir -p /tmp/data && cd /tmp/data
+docker run -v $PWD:/data vanessa/expfactory-builder-ci build test-task
 
 echo "Building container"
-cd /tmp/data
 docker build -t expfactory/experiments .
