+++ expfactory/templates/build/docker/builder-ci/build.sh
@@ -1,4 +1,4 @@
 #!/use/bin/bash
 
 echo "Building CI builder for testing current PR..."
-docker build --rm=false -t vanessa/expfactory-builder-ci .
+EXPFACTORY_BRANCH=$CIRCLE_BRANCH docker build --rm=false -t vanessa/expfactory-builder-ci .
