+++ expfactory/testing/test_builder.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 
 echo "Testing builder writing to /tmp"
+
 mkdir -p /tmp/data && cd /tmp/data
 docker run -v $PWD:/data vanessa/expfactory-builder-ci build test-task
 
