### 2017-12-15 00ef5c9
adding basic testing cases
