+++ expfactory/testing/data/test_task/js/jspsych/jspsych.js
@@ -1,1590 +1,0 @@
-/**
- * jspsych.js
- * Josh de Leeuw
- *
- * documentation: docs.jspsych.org
- *
- **/
-var jsPsych = (function() {
-
-  var core = {};
-
-  //
-  // private variables
-  //
-
-  // options
-  var opts = {};
-  // experiment timeline
-  var timeline;
-  // flow control
-  var global_trial_index = 0;
-  var current_trial = {};
-  // target DOM element
-  var DOM_target;
-  // time that the experiment began
-  var exp_start_time;
-
-  //
-  // public methods
-  //
-
-  core.init = function(options) {
-
-    // reset variables
-    timeline = null;
-    global_trial_index = 0;
-    current_trial = {};
-
-    // check if there is a body element on the page
-    var default_display_element = $('body');
-    if (default_display_element.length === 0) {
-      $(document.documentElement).append($('<body>'));
-      default_display_element = $('body');
-    }
-
-    var defaults = {
-      'display_element': default_display_element,
-      'on_finish': function(data) {
-        return undefined;
-      },
-      'on_trial_start': function() {
-        return undefined;
-      },
-      'on_trial_finish': function() {
-        return undefined;
-      },
-      'on_data_update': function(data) {
-        return undefined;
-      },
-      'show_progress_bar': false,
-      'auto_preload': true,
-      'max_load_time': 30000,
-      'skip_load_check': false,
-      'fullscreen': false,
-      'default_iti': 1000
-    };
-
-    // override default options if user specifies an option
-    opts = $.extend({}, defaults, options);
-
-    // set target
-    if (typeof window[opts.display_element] == 'function') {
-      DOM_target = window[opts.display_element]()
-    } else {
-      DOM_target = opts.display_element;
-    }    
-    
-    // add CSS class to DOM_target
-    DOM_target.addClass('jspsych-display-element');
-
-    // create experiment timeline
-    timeline = new TimelineNode({
-      timeline: opts.timeline
-    });
-
-    // preloading
-    if(opts.auto_preload){
-      jsPsych.pluginAPI.autoPreload(timeline, startExperiment);
-    } else {
-      startExperiment();
-    }
-  };
-
-  core.progress = function() {
-
-    var percent_complete = timeline.percentComplete()
-
-    var obj = {
-      "total_trials": timeline.length(),
-      "current_trial_global": global_trial_index,
-      "percent_complete": percent_complete
-    };
-
-    return obj;
-  };
-
-  core.startTime = function() {
-    return exp_start_time;
-  };
-
-  core.totalTime = function() {
-    return (new Date()).getTime() - exp_start_time.getTime();
-  };
-
-  core.getDisplayElement = function() {
-    return DOM_target;
-  };
-
-  core.finishTrial = function(data) {
-    // write the data from the trial
-    data = typeof data == 'undefined' ? {} : data;
-    jsPsych.data.write(data);
-
-    // get back the data with all of the defaults in
-    var trial_data = jsPsych.data.getDataByTrialIndex(global_trial_index);
-
-    // handle callback at plugin level
-    if (typeof current_trial.on_finish === 'function') {
-      current_trial.on_finish(trial_data);
-    }
-
-    // handle callback at whole-experiment level
-    opts.on_trial_finish(trial_data);
-
-    // wait for iti
-    if (typeof current_trial.timing_post_trial == 'undefined') {
-      if (opts.default_iti > 0) {
-        setTimeout(next_trial, opts.default_iti);
-      } else {
-        next_trial();
-      }
-    } else {
-      if (current_trial.timing_post_trial > 0) {
-        setTimeout(next_trial, current_trial.timing_post_trial);
-      } else {
-        next_trial();
-      }
-    }
[…]
