+++ expfactory/templates/build/docker/builder-base/README.md
@@ -1,0 +1,7 @@
+# Docker Builder Base
+
+This is the base image for the various builders
+
+```
+docker build -t vanessa/expfactory-builder-base .
+```
