+++ expfactory/templates/build/docker/builder-ci/Dockerfile
@@ -1,9 +1,6 @@
-FROM ubuntu:16.04
+FROM vanessa/expfactory-builder:base
 
 # docker build -t vanessa/expfactory-builder-ci .
-
-RUN apt-get update && apt-get install -y nginx git python3-pip python3-dev python apt-utils
-ENV DEBIAN_FRONTEND noninteractive
 
 ###################################
 # Experiment Factory
