### 2017-12-15 f25ebdb
modified:   circle.yml 	modified:   expfactory/templates/build/docker/Dockerfile.dev 	modified:   expfactory/templates/build/docker/Dockerfile.template
