### 2017-12-15 b48f639
adding users, logs command
