### 2017-12-15 2d85850
adding basic script to deploy docker, after merge
