### 2017-12-16 2282550
removing surveys generator from expfactory to be put in its own repository
