+++ examples/docker/custom-container/startscript.sh
@@ -1,47 +1,152 @@
 #!/bin/bash
 
+usage () {
+
+    echo "Usage:
+
+
+    
+         docker run <container> [help|list|test-experiments|start]
+         docker run -p 80:80 -v /tmp/data:/scif/data <container> start
+
+         Headless Mode (requires token)
+
+         docker run -p 80:80 -d --name experiments -v /tmp/data:/scif/data <container> --headless start
+         docker exec experiments expfactory users 3
+
+
+         Commands:
+
+                help: show help and exit
+                list: list installed experiments
+                lib: list experiments in the library
+                test: test experiments installed in container
+                start: start the container to do the experiments*
+                env: search for an environment variable set in the container
+         
+         *you are required to map port 80, otherwise you won't see the portal at localhost
+
+         Options [start]:
+
+                --db: specify a database url to override the default filesystem
+                                 [sqlite|mysql|postgresql]:///
+
+                --studyid:  specify a studyid to override the default
+                --randomize: select experiment order at random
+                --no-randomize: select experiment order manually in the GUI
+                --experiments: a comma separated list of experiments (manual ordering) 
+
+         Examples:
+
+              docker run <container> test
+              docker run <container> list
+              docker run <container> start
+              docker run -p 80:80 <container> --database mysql+pymysql://username:password@host/dbname start
+              docker run -p 80:80 <container> --database sqlite start
+              docker run -p 80:80 <container> --database postgresql://username:password@host/dbname start
+
+         "
+}
+
 if [ $# -eq 0 ]; then
-    echo "Usage:"
-    echo "docker run vanessa/expfactory-builder [help|list|test-experiments|start]"
-    echo "docker run -p 80:80 -v /tmp/data:/scif/data vanessa/expfactory-builder start"
+    usage
     exit
 fi
 
+EXPFACTORY_START="no"
+EXPFACTORY_DATABASE="filesystem"
 
 while true; do
     case ${1:-} in
         -h|--help|help)
-
-            echo "Additional commands:
-                  docker run vanessa/expfactory-builder test
-                  docker run vanessa/expfactory-builder list
-                  docker run vanessa/expfactory-builder start"
-
-            exec expfactory --help
+            usage
             exit
         ;;
-        -test-experiments|--te|test)
-            cd /opt/expfactory/expfactory/templates/build
-            exec python3 -m unittest tests.test_experiment
+        --database|--db)
+            shift
+            EXPFACTORY_DATABASE=${1:-}
+            export EXPFACTORY_DATABASE
+            shift
+        ;;
+        --env|env)
+            shift
+            env | grep ${1:-}
+            exit
+        ;;
+        -e|experiments|--experiments)
+            shift
+            EXPFACTORY_EXPERIMENTS="${1:-}"
+            shift
+            export EXPFACTORY_EXPERIMENTS
+        ;;
+        --headless|headless)
+            shift
+            EXPFACTORY_HEADLESS="true"
+            export EXPFACTORY_HEADLESS
+        ;;
+        --lib)
+            echo "Experiments in the library:"
+            expfactory list
+            echo
             exit
         ;;
         -ls|--list|list)
             echo "Experiments in this image:"
             ls /scif/apps -1
             echo
-            echo "Experiments in the library:"
-            exec expfactory list
             exit
         ;;
+        --studyid)
+            shift
+            EXPFACTORY_STUDY_ID=${1:-}
[…]
