+++ examples/docker/custom-container/README.md
@@ -1,3 +1,8 @@
 # Experiments Container
 
+```
+docker run -v $PWD:/data vanessa/expfactory-builder build digit-span spatial-span tower-of-london test-task
+docker build -t vanessa/experiment .
+```
+
 This container recipe was generated with the commands shown in [the generate docs](https://expfactory.github.io/expfactory/generate.html).  See them for building and usage.

[…]
