### 2017-12-18 #474 @leeper
Hi @nirum would you be interested in reviewing this for JOSS?
