+++ docs/_data/toc.yml
@@ -27,6 +27,13 @@
         slug: contribute-a-container
       - title: "Contribute an Experiment"
         slug: contribute-an-experiment
+
+  - title: "Integrations"
+    url: "/expfactory/integrations"
+    slug: integrations
+    sections:
       - title: "Contribute a Survey"
         slug: contribute-a-survey
+      - title: "Expfactory Robots"
+        slug: expfactory-robots
 
