+++ expfactory/templates/build/docker/builder-dev/Dockerfile
@@ -7,7 +7,7 @@ FROM vanessa/expfactory-builder:base
 ###################################
 
 WORKDIR /opt
-RUN git clone -b headless https://www.github.com/expfactory/expfactory
+RUN git clone -b add/labjs-support https://www.github.com/expfactory/expfactory
 WORKDIR /opt/expfactory
 RUN python3 setup.py install
 RUN python3 -m pip install pyaml
