### 2017-12-22 #474 @nirum
Hi all, looks interesting, but I don't have the bandwidth right now to give this a proper review :(
