### 2017-12-22 #474 @leeper
Hi @oesteban would you be interested in reviewing this for JOSS?
