### 2017-12-22 #474 @leeper
@oesteban That's great! Thank you! I've fired up a review ticket here: https://github.com/openjournals/joss-reviews/issues/521, which provides all instructions. Let me know if you have questions.
