### 2017-12-31 expfactory/expfactory#54 @oesteban
[JOSS Review] Community guidelines
As per openjournals/joss-reviews#521:

> **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I think the project would greatly benefit if:

  - Added the GitHub's magics for contributions (https://github.com/expfactory/expfactory/community)
  - The README was updated giving very brief (but present) pointers to how to communicate.
