### 2017-12-31 expfactory/expfactory#55 @oesteban
[JOSS Review] Installation instructions
As per openjournals/joss-reviews#521:

> **Installation instructions**: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Even though the project is well documented, I have the impression that the information could be very unclear to someone who does not know about docker or containers. I like the approach of let's dive in ASAP, but for unexperienced users that likely will hold them off jumping. Especially when other barriers come into play (e.g. you are not English native speaker, like myself), the documentation might be a bit overwhelming/discouraging.

A separate section where the documentation is more "traditional" (with clear steps and sections) would be very much appreciated. Or maybe a video clip showing how to run the containers *a la* datalad?

I don't think this is a major issue though, just trying to give a different point of view.
