### 2017-12-31 c6bf2fa
adding contributions files to close #54
