+++ .github/ISSUE_TEMPLATE.md
@@ -1,0 +1,16 @@
+## Version of Experiment Factory:
+
+Write here.
+
+### Expected behavior
+
+Write here.
+
+### Actual behavior
+
+Write here.
+
+### Steps to reproduce behavior
+
+Write here.
+
