### 2017-12-31 expfactory/expfactory#54 @vsoch
Thanks @oesteban ! This is very good and sound advice, and I've added community guidelines (detailed in the README) to address the point. If you want to take a look --> https://github.com/expfactory/expfactory/pull/56 and please let me know if any language should be changed, or anything made better or more clear.
