### 2017-12-31 expfactory/expfactory#55 @vsoch
hey @oesteban! This is a good point - the first set of instructions assume that the user understand / knows Docker, and if this isn't the case, it's a hard entry. I'm going to try adding a more direct link from the first page and the README to this page --> https://expfactory.github.io/expfactory/generate#detailed-start and then I can write a bit more about containers there. I'll add this to the current PR and let you know when it's ready for your review.
