### 2017-12-31 03b8051
adding background and links to it to docs (for a gentler start)
