### 2017-12-31 expfactory/expfactory#55 @vsoch
okay, I've added a gentle introduction to both (containers and then using expfactory) to the PR. I intentionally have not shown the user how to install expfactory on the host because I want the primarily interaction to be via a builder container, so I am not showing the traditional `pip install` as you would expect with other Python software. In that the container is provided, the interaction with the software is best done via the container so users avoid the dependency issues that arise when trying to install native to the host.
