### 2018-01-01 expfactory/expfactory#55 @earcanal
I would remove references to Singularity for now, or explain _very clearly_ how it fits into the (container) picture.  My understanding is that Docker is more stable than Singularity?  The consequence of that for me was that my understanding increased rapidly once I got my hands a bit grubby with Docker.
