### 2018-01-01 f7e9a63
removing references to singularity where not relevant at suggestion of @earcanal
