+++ setup.py
@@ -7,9 +7,9 @@ import os
 ##########################################################################################
 
 def get_lookup():
-    '''get version by way of singularity.version, returns a 
+    '''get version by way of expfactory.version, returns a 
     lookup dictionary with several global variables without
-    needing to import singularity
+    needing to import expfactory
     '''
     lookup = dict()
     version_file = os.path.join('expfactory', 'version.py')
