### 2018-01-01 expfactory/expfactory#55 @vsoch
good call @earcanal ! This is an artifact of the fact that I first implemented this entire thing with Singularity, and then changed my mind, haha. I agree that Docker turned out to be better suited, as it's optimized for web services (and this functionality is only starting development for Singularity). I updated the PR to remove these stale references. The only remaining are with regards to integrations (because there are Singularity containers for some of the other tools provided).
