### 2018-01-09 a6566ac
Add ReverseProxied middleware.
Flask snippet 35: http://flask.pocoo.org/snippets/35/
