### 2018-01-09 979710f
Change redirects to use url_for() so that app works behind non-root URL proxy.
