### 2018-01-13 5a88b40
Return correct ajax responses in /save.
Was a Typo: 'ContentType'.
