+++ expfactory/views/main.py
@@ -122,9 +122,8 @@ def save():
         app.logger.info('Finished %s, %s remaining.' % (exp_id, len(experiments)))
 
         # Note, this doesn't seem to be enough to trigger ajax success
-        return json.dumps({'success':True}), 200, {'ContentType':'application/json'} 
-    return json.dumps({'success':False}), 403, {'ContentType':'application/json'} 
-
+        return jsonify(success=True)
+    return jsonify(success=False)
 
 
 @app.route('/next', methods=['POST', 'GET'])
