### 2018-01-13 48dc90e
Insert experiment variables (`csrf_token` and `SCRIPT_ROOT`) template into experiment index.html on install.
Previous solution of appending the template script after the html tag
was unreliable.

SCRIPT_ROOT is necessary for posting to experiment factory deployed
behind a reverse proxy. Experiment must POST to SCRIPT_ROOT + '/save'.

```js
on_finish: function(data) {
  // Get the data.
  var data = jsPsych.data.get();

  var request = $.post(
    SCRIPT_ROOT + '/save',
    {'data': data.json()},
    function() {
      // Success, go to /next.
      document.location = SCRIPT_ROOT + '/next';
  
[…]
