### 2018-01-13 52b41af
Add beautifulsoup4 to dockerfiles.
