### 2018-01-31 expfactory/expfactory#55 @vsoch
Contributions are merged, thanks for the help!
