### 2018-02-04 #521 @leeper
@oesteban Can you confirm that you are happy with the package in its current form?
